requirements.txt
python-telegram-bot>=21.0,<22
anthropic>=0.25.0
fpdf2>=2.7.0
python-dotenv>=1.0.0
